finally working version with updating sorting numbers on all items in topic/template

+++ src/resolvers/types/QuestionTemplateRelation.ts
@@ -1,4 +1,4 @@
-import { Field, Float, Int, ObjectType } from 'type-graphql';
+import { Field, Int, ObjectType } from 'type-graphql';
 
 import { QuestionTemplateRelation as QuestionTemplateRelationOrign } from '../../models/Template';
 import { FieldConfigType } from './FieldConfig';
@@ -11,7 +11,7 @@ export class QuestionTemplateRelation
   @Field(() => Question)
   public question: Question;
 
-  @Field(() => Float)
+  @Field(() => Int)
   public sortOrder: number;
 
   @Field(() => Int)
